+++ lib/uuid.dart
@@ -3,6 +3,8 @@ library uuid;
 
 import 'dart:typed_data';
 
+import 'package:uuid/uuid_value.dart';
+
 import 'uuid_util.dart';
 import 'package:crypto/crypto.dart' as crypto;
 
@@ -23,54 +25,46 @@ class Uuid {
     return i.toRadixString(16).padLeft(2, '0');
   });
 
-  var _seedBytes, _nodeId, _clockSeq, _lastMSecs = 0, _lastNSecs = 0;
-  late List<int> Function() _globalRNG;
+  final options;
+  static final _state = {
+    'seedBytes': null,
+    'node': null,
+    'clockSeq': null,
+    'mSecs': 0,
+    'nSecs': 0,
+    'hasInitV1': false,
+    'hasInitV4': false
+  };
 
-  Uuid({Map<String, dynamic>? options}) {
-    options ??= const {};
+  const Uuid({Map<String, dynamic>? this.options});
 
-    // Sets initial seedBytes, node, and clock seq based on mathRNG.
-    var v1PositionalArgs = (options['v1rngPositionalArgs'] != null)
-        ? options['v1rngPositionalArgs']
-        : [];
-    var v1NamedArgs = (options['v1rngNamedArgs'] != null)
-        ? options['v1rngNamedArgs'] as Map<Symbol, dynamic>
-        : const <Symbol, dynamic>{};
-    _seedBytes = (options['v1rng'] != null)
-        ? Function.apply(options['v1rng'], v1PositionalArgs, v1NamedArgs)
-        : UuidUtil.mathRNG();
+  static bool isValidUUID(String uuid) {
+    // UUID of all 0s is ok.
+    if (uuid == NAMESPACE_NIL) {
+      return true;
+    }
 
-    // Set the globalRNG function to mathRNG with the option to set an alternative globally
-    var gPositionalArgs = (options['grngPositionalArgs'] != null)
-        ? options['grngPositionalArgs']
-        : [];
-    var gNamedArgs = (options['grngNamedArgs'] != null)
-        ? options['grngNamedArgs'] as Map<Symbol, dynamic>
-        : const <Symbol, dynamic>{};
+    // If its not 36 characters in length, don't bother (including dashes).
+    if (uuid.length != 36) {
+      return false;
+    }
 
-    final grng = options['grng'];
-    _globalRNG = (grng != null)
-        ? () => Function.apply(grng, gPositionalArgs, gNamedArgs)
-        : UuidUtil.mathRNG;
-
-    // Per 4.5, create a 48-bit node id (47 random bits + multicast bit = 1)
-    _nodeId = [
-      _seedBytes[0] | 0x01,
-      _seedBytes[1],
-      _seedBytes[2],
-      _seedBytes[3],
-      _seedBytes[4],
-      _seedBytes[5]
-    ];
-
-    // Per 4.2.2, randomize (14 bit) clockseq
-    _clockSeq = (_seedBytes[6] << 8 | _seedBytes[7]) & 0x3ffff;
+    // Make sure if it passes the above, that its valid.
+    const pattern =
+        r'^[0-9a-f]{8}-[0-9a-f]{4}-[1-5][0-9a-f]{3}-[89ab][0-9a-f]{3}-[0-9a-f]{12}$';
+    final regex = RegExp(pattern, caseSensitive: false, multiLine: true);
+    final match = regex.hasMatch(uuid);
+    return match;
   }
 
   ///Parses the provided [uuid] into a list of byte values.
   /// Can optionally be provided a [buffer] to write into and
   ///  a positional [offset] for where to start inputting into the buffer.
-  List<int> parse(String uuid, {List<int>? buffer, int offset = 0}) {
+  /// Throws FormatException if the UUID is invalid.
+  static List<int> parse(String uuid, {List<int>? buffer, int offset = 0}) {
+    if (!isValidUUID(uuid)) {
+      throw FormatException('The provided UUID is invalid.', uuid);
+    }
     var i = offset, ii = 0;
 
     // Create a 16 item buffer if one hasn't been provided.
@@ -98,7 +92,11 @@ class Uuid {
   /// Unparses a [buffer] of bytes and outputs a proper UUID string.
   /// An optional [offset] is allowed if you want to start at a different point
   /// in the buffer.
-  String unparse(List<int> buffer, {int offset = 0}) {
+  /// Throws an exception if the buffer does not have a length of 16
+  static String unparse(List<int> buffer, {int offset = 0}) {
+    if (buffer.length != 16) {
+      throw Exception('The provided buffer needs to have a length of 16.');
+    }
     var i = offset;
     return '${_byteToHex[buffer[i++]]}${_byteToHex[buffer[i++]]}'
         '${_byteToHex[buffer[i++]]}${_byteToHex[buffer[i++]]}-'
@@ -108,6 +106,66 @@ class Uuid {
         '${_byteToHex[buffer[i++]]}${_byteToHex[buffer[i++]]}'
         '${_byteToHex[buffer[i++]]}${_byteToHex[buffer[i++]]}'
         '${_byteToHex[buffer[i++]]}${_byteToHex[buffer[i++]]}';
+  }
+
+  void _initV1(Map<String, dynamic>? options) {
+    options ??= const {};
+
+    if (!(_state['hasInitV1']! as bool)) {
+      var v1PositionalArgs = (options['v1rngPositionalArgs'] != null)
+          ? options['v1rngPositionalArgs']
+          : [];
+      var v1NamedArgs = (options['v1rngNamedArgs'] != null)
+          ? options['v1rngNamedArgs'] as Map<Symbol, dynamic>
+          : const <Symbol, dynamic>{};
+      Uint8List seedBytes = (options['v1rng'] != null)
+          ? Function.apply(options['v1rng'], v1PositionalArgs, v1NamedArgs)
+          : UuidUtil.mathRNG();
+
+      (_state['seedBytes'] != null)
+          ? _state['seedBytes']
+          : _state['seedBytes'] = seedBytes;
+
+      // Per 4.5, create a 48-bit node id (47 random bits + multicast bit = 1)
+      var nodeId = [
+        seedBytes[0] | 0x01,
+        seedBytes[1],
+        seedBytes[2],
+        seedBytes[3],
+        seedBytes[4],
+        seedBytes[5]
+      ];
+      (_state['node'] != null) ? _state['node'] : _state['node'] = nodeId;
+
+      // Per 4.2.2, randomize (14 bit) clockseq
+      var clockSeq = (seedBytes[6] << 8 | seedBytes[7]) & 0x3ffff;
+      _state['clockSeq'] ??= clockSeq;
+
+      _state['mSecs'] = 0;
+      _state['nSecs'] = 0;
+      _state['hasInitV1'] = true;
+    }
+  }
+
+  void _initV4(Map<String, dynamic>? options) {
+    options ??= {};
+
+    if (!(_state['hasInitV4']! as bool)) {
+      // Set the globalRNG function to mathRNG with the option to set an alternative globally
+      var gPositionalArgs = (options['grngPositionalArgs'] != null)
+          ? options['grngPositionalArgs']
+          : [];
+      var gNamedArgs = (options['grngNamedArgs'] != null)
+          ? options['grngNamedArgs'] as Map<Symbol, dynamic>
+          : const <Symbol, dynamic>{};
+
+      final grng = options['grng'];
+      _state['globalRNG'] = (grng != null)
+          ? () => Function.apply(grng, gPositionalArgs, gNamedArgs)
+          : UuidUtil.mathRNG;
+
+      _state['hasInitV4'] = true;
+    }
   }
 
   /// v1() Generates a time-based version 1 UUID
@@ -122,10 +180,12 @@ class Uuid {
   String v1({Map<String, dynamic>? options}) {
     var i = 0;
     var buf = Uint8List(16);
-    options = (options != null) ? options : {};
+    options ??= {};
 
-    var clockSeq =
-        (options['clockSeq'] != null) ? options['clockSeq'] : _clockSeq;
+    _initV1(options);
+    var clockSeq = (options['clockSeq'] != null)
+        ? options['clockSeq']
+        : _state['clockSeq'] as int;
 
     // UUID timestamps are 100 nano-second units since the Gregorian epoch,
     // (1582-10-15 00:00). Time is handled internally as 'msecs' (integer
@@ -137,10 +197,12 @@ class Uuid {
 
     // Per 4.2.1.2, use count of uuid's generated during the current clock
     // cycle to simulate higher resolution clock
-    var nSecs = (options['nSecs'] != null) ? options['nSecs'] : _lastNSecs + 1;
+    var nSecs = (options['nSecs'] != null)
+        ? options['nSecs']
+        : (_state['nSecs']! as int) + 1;
 
     // Time since last uuid creation (in msecs)
-    var dt = (mSecs - _lastMSecs) + (nSecs - _lastNSecs) / 10000;
+    var dt = (mSecs - _state['mSecs']) + (nSecs - _state['nSecs']) / 10000;
 
     // Per 4.2.1.2, Bump clockseq on clock regression
     if (dt < 0 && options['clockSeq'] == null) {
@@ -149,7 +211,7 @@ class Uuid {
 
     // Reset nsecs if clock regresses (new clockseq) or we've moved onto a new
     // time interval
-    if ((dt < 0 || mSecs > _lastMSecs) && options['nSecs'] == null) {
+    if ((dt < 0 || mSecs > _state['mSecs']) && options['nSecs'] == null) {
       nSecs = 0;
     }
 
@@ -158,9 +220,9 @@ class Uuid {
       throw Exception('uuid.v1(): Can\'t create more than 10M uuids/sec');
     }
 
-    _lastMSecs = mSecs;
-    _lastNSecs = nSecs;
-    _clockSeq = clockSeq;
+    _state['mSecs'] = mSecs;
+    _state['nSecs'] = nSecs;
+    _state['clockSeq'] = clockSeq;
 
     // Per 4.1.4 - Convert from unix epoch to Gregorian epoch
     mSecs += 12219292800000;
@@ -188,7 +250,7 @@ class Uuid {
     buf[i++] = clockSeq & 0xff;
 
     // node
-    var node = (options['node'] != null) ? options['node'] : _nodeId;
+    var node = (options['node'] != null) ? options['node'] : _state['node'];
     for (var n = 0; n < 6; n++) {
       buf[i + n] = node[n];
     }
@@ -215,6 +277,20 @@ class Uuid {
     return parse(v1(options: options), buffer: buffer, offset: offset);
   }
 
+  /// v1obj() Generates a time-based version 1 UUID
+  ///
+  /// By default it will generate a string based off current time, and will
+  /// return it as a [UuidValue] object.
+  ///
+  /// The first argument is an options map that takes various configuration
+  /// options detailed in the readme.
+  ///
+  /// http://tools.ietf.org/html/rfc4122.html#section-4.2.2
+  UuidValue v1obj({Map<String, dynamic>? options}) {
+    var uuid = v1(options: options);
+    return UuidValue(uuid);
+  }
+
   /// v4() Generates a RNG version 4 UUID
   ///
   /// By default it will generate a string based mathRNG, and will return
@@ -227,6 +303,7 @@ class Uuid {
   String v4({Map<String, dynamic>? options}) {
     options = (options != null) ? options : <String, dynamic>{};
 
+    _initV4(options);
     // Use the built-in RNG or a custom provided RNG
     var positionalArgs =
         (options['positionalArgs'] != null) ? options['positionalArgs'] : [];
@@ -235,7 +312,7 @@ class Uuid {
         : const <Symbol, dynamic>{};
     var rng = (options['rng'] != null)
         ? Function.apply(options['rng'], positionalArgs, namedArgs)
-        : _globalRNG();
+        : (_state['globalRNG']! as Function)();
 
     // Use provided values over RNG
     var rnds = (options['random'] != null) ? options['random'] : rng;
@@ -265,6 +342,20 @@ class Uuid {
     int offset = 0,
   }) {
     return parse(v4(options: options), buffer: buffer, offset: offset);
+  }
+
+  /// v4obj() Generates a RNG version 4 UUID
+  ///
+  /// By default it will generate a string based mathRNG, and will return
+  /// a [UuidValue] object. If you wish to use crypto-strong RNG, pass in UuidUtil.cryptoRNG
+  ///
+  /// The first argument is an options map that takes various configuration
+  /// options detailed in the readme.
+  ///
+  /// http://tools.ietf.org/html/rfc4122.html#section-4.4
+  UuidValue v4obj({Map<String, dynamic>? options}) {
+    var uuid = v4(options: options);
+    return UuidValue(uuid);
   }
 
   /// v5() Generates a namspace & name-based version 5 UUID
@@ -309,7 +400,7 @@ class Uuid {
     hashBytes[6] = (hashBytes[6] & 0x0f) | 0x50;
     hashBytes[8] = (hashBytes[8] & 0x3f) | 0x80;
 
-    return unparse(hashBytes);
+    return unparse(hashBytes.sublist(0, 16));
   }
 
   /// v5buffer() Generates a RNG version 4 UUID
@@ -332,5 +423,20 @@ class Uuid {
   }) {
     return parse(v5(namespace, name, options: options),
         buffer: buffer, offset: offset);
+  }
+
+  /// v5obj() Generates a namspace & name-based version 5 UUID
+  ///
+  /// By default it will generate a string based on a provided uuid namespace and
+  /// name, and will return a [UuidValue] object.
+  ///
+  /// The first argument is an options map that takes various configuration
+  /// options detailed in the readme.
+  ///
+  /// http://tools.ietf.org/html/rfc4122.html#section-4.4
+  UuidValue v5obj(String? namespace, String? name,
+      {Map<String, dynamic>? options}) {
+    var uuid = v5(namespace, name, options: options);
+    return UuidValue(uuid);
   }
 }